Rename rebuild_stats files
