+++ packages/devtools_app/lib/src/screens/performance/panes/frame_analysis/frame_analysis.dart
@@ -12,8 +12,8 @@ import '../../../../shared/globals.dart';
 import '../../../../shared/theme.dart';
 import '../../../../ui/colors.dart';
 import '../controls/enhance_tracing/enhance_tracing_controller.dart';
-import '../rebuild_stats/rebuild_counts.dart';
-import '../rebuild_stats/widget_rebuild.dart';
+import '../rebuild_stats/rebuild_stats_model.dart';
+import '../rebuild_stats/rebuild_stats.dart';
 import 'frame_analysis_model.dart';
 import 'frame_hints.dart';
 import 'frame_time_visualizer.dart';
