+++ packages/devtools_app/lib/src/screens/performance/panes/rebuild_stats/rebuild_stats_model.dart
