+++ packages/devtools_app/test/performance/frame_analysis/frame_analysis_test.dart
@@ -7,7 +7,7 @@ import 'package:devtools_app/src/config_specific/import_export/import_export.dar
 import 'package:devtools_app/src/screens/performance/panes/frame_analysis/frame_analysis.dart';
 import 'package:devtools_app/src/screens/performance/panes/frame_analysis/frame_hints.dart';
 import 'package:devtools_app/src/screens/performance/panes/frame_analysis/frame_time_visualizer.dart';
-import 'package:devtools_app/src/screens/performance/panes/rebuild_stats/rebuild_counts.dart';
+import 'package:devtools_app/src/screens/performance/panes/rebuild_stats/rebuild_stats_model.dart';
 import 'package:devtools_test/devtools_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
