+++ packages/devtools_app/test/performance/performance_model_test.dart
@@ -6,7 +6,7 @@ import 'dart:convert';
 
 import 'package:devtools_app/devtools_app.dart';
 import 'package:devtools_app/src/screens/performance/panes/raster_stats/raster_stats_model.dart';
-import 'package:devtools_app/src/screens/performance/panes/rebuild_stats/rebuild_counts.dart';
+import 'package:devtools_app/src/screens/performance/panes/rebuild_stats/rebuild_stats_model.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../test_infra/test_data/cpu_profile.dart';
